+++ bull-common/src/test/java/com/expediagroup/beans/sample/ToFooSimpleNestedWithY.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2018-2023 Expedia, Inc.
+ * Copyright (C) 2019-2026 Expedia, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package com.expediagroup.beans.sample;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.ToString;
+
 /**
- * Example usage of the bytecode adapter plus other sample types used in tests.
+ * Nested destination bean with a differently named field (y instead of x).
+ * Used to test that breadcrumb-based mapping correctly resolves when source and destination field names differ.
  */
-package com.expediagroup.beans.generator.bytecode.sample;
+@AllArgsConstructor
+@Getter
+@ToString
+public class ToFooSimpleNestedWithY {
+    private final String name;
+    private final long index;
+    private final int y;
+}